pyproject.toml
[tool.poetry]
name = "telegram-bot-local-dev"
version = "0.1.0"
description = "Local development template for Telegram bots with python-telegram-bot and aiogram"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "ptb"},
    {include = "aiogram"},
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
python-telegram-bot = "^21.6"
aiogram = "^3.10.0"
aiohttp = "^3.9.5"
websockets = "^12.0"
httpx = "^0.27.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
structlog = "^24.2.0"
tenacity = "^8.5.0"
cachetools = "^5.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
mypy = "^1.8.0"
types-python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Мягкая настройка для начала
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "telegram.*",
    "aiogram.*",
]
ignore_missing_imports = true
